+++ client/operations/put_porosity_simulation_parameters.go
@@ -63,7 +63,7 @@ type PutPorositySimulationParams struct {
 	  simulation identifier
 
 	*/
-	ID int64
+	ID int32
 
 	timeout    time.Duration
 	Context    context.Context
@@ -104,13 +104,13 @@ func (o *PutPorositySimulationParams) SetPorositySimulation(porositySimulation *
 }
 
 // WithID adds the id to the put porosity simulation params
-func (o *PutPorositySimulationParams) WithID(id int64) *PutPorositySimulationParams {
+func (o *PutPorositySimulationParams) WithID(id int32) *PutPorositySimulationParams {
 	o.SetID(id)
 	return o
 }
 
 // SetID adds the id to the put porosity simulation params
-func (o *PutPorositySimulationParams) SetID(id int64) {
+func (o *PutPorositySimulationParams) SetID(id int32) {
 	o.ID = id
 }
 
@@ -129,7 +129,7 @@ func (o *PutPorositySimulationParams) WriteToRequest(r runtime.ClientRequest, re
 	}
 
 	// path param id
-	if err := r.SetPathParam("id", swag.FormatInt64(o.ID)); err != nil {
+	if err := r.SetPathParam("id", swag.FormatInt32(o.ID)); err != nil {
 		return err
 	}
