+++ client/operations/start_scan_pattern_simulation_parameters.go
@@ -56,7 +56,7 @@ type StartScanPatternSimulationParams struct {
 	  simulation identifier
 
 	*/
-	ID int64
+	ID int32
 
 	timeout    time.Duration
 	Context    context.Context
@@ -86,13 +86,13 @@ func (o *StartScanPatternSimulationParams) SetContext(ctx context.Context) {
 }
 
 // WithID adds the id to the start scan pattern simulation params
-func (o *StartScanPatternSimulationParams) WithID(id int64) *StartScanPatternSimulationParams {
+func (o *StartScanPatternSimulationParams) WithID(id int32) *StartScanPatternSimulationParams {
 	o.SetID(id)
 	return o
 }
 
 // SetID adds the id to the start scan pattern simulation params
-func (o *StartScanPatternSimulationParams) SetID(id int64) {
+func (o *StartScanPatternSimulationParams) SetID(id int32) {
 	o.ID = id
 }
 
@@ -103,7 +103,7 @@ func (o *StartScanPatternSimulationParams) WriteToRequest(r runtime.ClientReques
 	var res []error
 
 	// path param id
-	if err := r.SetPathParam("id", swag.FormatInt64(o.ID)); err != nil {
+	if err := r.SetPathParam("id", swag.FormatInt32(o.ID)); err != nil {
 		return err
 	}
